Implement Edit Positon multi-dropping
Right-clicking a piece in Edit Position mode now 'lifts it off the board',
and makes it available for dropping multiple times, by left-clicking
empty squares while no piece is selected (highlighted) for moving.
Left-clicks on occupied squares select the piece for moving as usual,
but this does not erase the memory of the 'lifted' piece types. This way
click-click moves remain possible in between drop events, e.g. to get a
piece out of the way.
 The type of the piece that will be dropped is indicated in the message
field. Right-clicking an empty square (which is an invalid action, as empty
squares cannot be lifted for dropping) pops up a note with an overview of
what can be done in Edit Position mode.
 This is an alternative for sweep selection as a method of dropping pieces
on the board. For now it replaces the piece menu. Multiple drop clicks to
the same square will increment the piece type, so piece types not in the
original position can still be created. It doesn't seem possible to do
anything with a board that is completely empty, though!