Implement auto-promotion
A piece that in the pieceToCharTable is marked as a promoted version of
some other piece by using a '-' instead of a '^' will now define both
partners as auto-promoting. This means promotion suffixes on SAN moves
for them will be suppressed an ignored, while in protocol moves they will
always be added. It will also be allowed to drop these pieces in promoted
form. The notation for this will use the nickname of the promoted version,
rather than a +X notation.