Finish change in Clear Board
Clock clicks in Edit Position mode now first displays a 'palette board',
then clears any non-Kings, then sets up the start position, and finally
restores the position that was erased by the palette board (if that was
different from the start position). This continues cyclically.
  The clearing of non-royals spares the four center squares, which would
have been empty already in an unmodified palette board. This allows a user
that wants to set up a position with very few pieces to first drag those
to the center, before clearing all other non-royals away, and finally
moving what is left to the desired location. That saves him the trouble
of eliminating the unneeded pieces one by one from the pallette board.