Fix echoing of ICS commands in ICS/Chat window
The ICS output pane did not show the commands submitted to the ICS
by typing in the input field. Now it does, but to prevent the password
from being shown this is made dependent on the value of 'remoteEchoOption'.
Which, to this end, had to be made an external variable shared through
backend.h.