Pay attention to hint move in walking pobder PV
When requesting to walk the PV of a pondering engine, it would only
work if that engine had printed the ponder move (in parentheses) as
first move of the PV. But the recommended mode in CECP would have
provided the ponder move in a 'Hint:' command, and start the PV after it.
So that the PV would always start with an invalid move (for the other
player). Backing up to the previous position such as would have to be
done when walking an old PV after the engine has moved, does not solve
this; we have to go one position forward rather than backward. Now
we first try to put the hint move before the PV, if the engine supplying
the PV provided the hint, and pondering is on, before we try to back up.