Do not nice nice-aware adapters
It is the engine process we want to run at lower priority, not the
adapter it is running through. So if we know the engine is running
through an adapter, because we invoked it ourselves due to -f/sUCI,
and we pass the adapter the value of the -niceEngines option (presumably
for it to set the priority of the engine it forks off by itself),
we refrain from nicing it.