Refrain from killing known protocol adapters
Killing 'engines' that run through an adapter and won't terminate fast
enough for XBoard's taste will always be very detrimental, as it will
kill the adapter, and leaves the actual engine (for which the adapter
was likely waiting) running, preventing the adapter would take action
against it. We will thus always refrain from killing engines when XBoard
knows they are running through an adapter, because it invoked the adapter
itself in response to the -f/sUCI option.