Interpret UCI2WB's invalid pieceToCharTable
UCI2WB sends an invalid pieceToCharTable "-" in its setup command for
engine-defined variants. This would keep the pieceToChar assignment
of the parent variant (in UCI2WB's case always 'fairy'). Now this is
replaced by the pieceToCharTable for variant normal, as the UCI engines
that use this feature only support variants with orthodox pieces.