Make square-size rounding dependent on fairy pieces (WB)
WinBoard 'rounds' the square size to one of the three values for which
images of fairy pieces are built in, in variants that normally would
contain such pieces. But when used as parent of an engine-defined variant
they might actually not contain any fairy pieces at all. This situation
is now recognized for parent variant fairy, and would suppress the rounding
when it occurs.