Improve XBetza e.p. system
The 'n' modifier can now be used on non-final legs of a multi-leg move
to indicate the target square of that leg becomes e.p. square. Pieces
that define an XBetza 'e' move will be able to use that move to e.p. capture
the last-moved piece when that moves goes to the e.p. square. To implement
this a routine was added to test whether this is the case (or whether the
e.p. capable piece used another move to go to the e.p. square).
  The limitation that there can only be more than a single e.p. square
when there is a pair of vertically adjacent ones still applies.