Support preludes in human-engine games
In an engine-defined variant the engine can now show several boards at
the start of the game by sending those to the GUI in a 'setup' command,
by letting all non-final 'setup' commands specify the (newly introduced)
parent variant 'prelude'. These board are intended as 'graphical menus',
from which the user can select a piece or square by clicking on it.
The 'lift' command of the highlight protocol can be used to relay such
clicks to the engine, and prompt it for the next 'setup' command.