Implement auto-gating in holdingless S-Chess
When there are no holdings in (parent) variant seirawan, the first and
last rank are supposed to act as file-specific holdings, and pieces
on it will automatically be gated if the virgin piece standing before
them is moved away by the user. (Engines must still specify a gating
through the promotion suffix.) The old location of the gated piece
on the holdings rank will then be blacked out (and the unused squares
on the holdings raks should also be black).