Suppress testing of holdings in holdingless seirawan
The legality test in variant seirawan normally tests if a piece specified
for gating is indeed in hand. This is now suppressed if there are no
holdings. There is no check on the correctness of the indicated gating
piece at all now, in that case. But presence of the suffix is still
necessary to trigger the gating of the piece that should be gated.