Make gating mandatory in holdingless Seirawan
Moving a piece 'in front of' a gatable piece is made mandatory,
by supplying the missing gating suffix in the move parser. To
achieve this the move type returned by LegalityTast() for a move
with missing suffix is White/BlackPromotion, even when the move is
a castling. The parser then adds the suffix, like it does for any
promotion without a suffix.