Fix matches from set-up position in engine-defined variants
Unfortunately XBoard loads positions from a position file before it
starts the engine for the next game in a match. For an engine-defined
game this means the engine did not get the opportunity to define a
pieceToCharTable with its 'setup'  command before the FEN is parsed,
with as a result that non-standard pieces in the FEN will not be
recognized. This is now fixed in a kludgy way, by remembering the
pieceToCharTable an engine sends as a first response to the 'variant'
command, and using that for the next game in a match as well, before
attempting to parse the FEN.