Refactor clock display code
The functions DisplayTimerLabel() in XBoard and DisplayAClock() in
WinBoard now first create the text to display in a buffer, with the
possible line break in logo mode indicated as underscore. This
underscore is then later replaced by linefeed or space, depending
on whether we want a one-line or two-line clock. This as preparation
for also printing other texts.