Allow entry of shogi promotions in any variant
Defining pieces as shogi-promotion partners did not cause triggering
of the promotion procedure when the corresponding basic piece reached
the zone for variants other than (Chu) Shogi. Now HasPromotionChoice()
recognizes pieces with a partner with ID '+' as promoting in any variant.
In variant fairy promotion on leaving the zone is now suppressed.
(This used to be only the case in variant elven.) This allows two-stage
promotions, where a Pawn first promote to a basic piece, and after leaving
the zone, the basic piece can promote again to its fixed partner.
(Like in Chess and a Half.)