Refactor ApplyMove
Almost all the cases in the large if-then-else sequence for the various
special moves examined the from-square and the set it to empty.
While a copy of the from-square was already in the variable 'piece'.
We now moved setting piece to the very beginning of the handling of
non-drop moves, and clear the from-square immediately after that,
instead of all the cases separately.