Implement XBetza unload modifier
Move legs containing a 'u' modifier will cause a piece captured
or destroyed by the entire move to reappear on the starting square of
that leg. This allows specification of swappers and pieces that push
other pieces elsewhere. The encoding of the corresponding moves uses
the same format as that for locust capture, and ApplyMove() distinguishes
those from the occupation of the specified 'kill square'. For now
this can only be used in a move without other (true) locust capture.
Notation (SAN), and reading back of such moves is still unsolved.