Fix oblique multi-leg moves in XBetza
Continuation legs of oblique atoms have thei directional modifiers
interpreted in the K system. But this had as an intended side effect
that they were split into a diagonal and orthogonal atom for separate
treatment, which for obliques is of course not possible, nor needed.
So this code is bypassed for the obliques now.