Fix PV conversion to SAN during pondering
Converting ponder PVs to SAN did not work when the ponder move was
given through a Hint command, rather than prefixed to the PV is
parentheses. Also, during TwoMachines mode, the conversion was not
invoked at all during pondering (as PV handling is done by another
code section in that case).