Implement Janggi as standard variant
The rules of Janggi (Korean Chess) are so peculiar amd unique that
it doesn't make sense to support them in terms of general applicable
features that can be configured by the engine. So the rules are now
hard-coded, and a new standard variant 'janggi' is added to select them.
This commit just takes care of the piece moves and initial position.