Fix spurious promotion for redefined Pawns in Xiangqi
The Betza move generator, used for pieces with moves redefined by the
'piece' command, always generated Pawn moves to last rank as promotions.
In Xiangqi this had the undesired effect that promotion suffixes were
considered missing on such moves arriving as text, and a default promotion
to Queen was appended to them. Which made the Pawns disappear, as a Queen
normally does not participate in Xiangqi.
  Now the MovesFromString routine suppresses promotion in Xiangqi and
Janggi. It still doesn't pay attention to zone depth in other variants.