Implement symmetrized books
Variants with a symmetric setup, (such as Xiangqi and Janggi) only need
to store half the positions. This is now implemented for books with
a filename *-sym.bin. For each position a 'canonical representation'
is defined, which is either the position itself or its mirror image,
depending on the first left-right difference. The book is then probed
for the cannonical representation, and the retreived moves are flipped
to the original representation when needed.