Redesign New Variant dialog of WinBoard
Because of the large number of engine-defined variants supported
by some engines, radiobuttons are no longer a suitable interface
for selecting the variant. These are now replaced by a single combobox,
which is filled with all standard variants in -ncp mode, and with
all variants (standard or engine-defined) supported by the first engine
if an engine is loaded.