Make some new General Options translatable in WinBoard
Some of the new checkboxes had not been added to the list of translatable
items yet. The radiobuttons in the New Variant dialog no longer exist,
so these were removed from the list. Instead the names of the standard
variants now are subjected to translation when they are added to the
variant-selection combobox.