Update JAWS patch for version 4.9
The JAWS patch was still at the stage of 4.5. So many piece names
had to be added to expand from 22 to 66 types. To prevent having
too many versions of the same name, the routine PieceToName() is
now adding the color prefixes, rather than taking them from the name table.
Calls to GenLegal() and HAsPromotionChoice() need an extra argument
in 4.9.