Delay reciting of requested piece type in JAWS version
When the Move Type-In dialog closes after flushing its speech output
the NVDA screen reader prefers to read the title bar of the main window,
aborting the desired output. This has now been cured by flushing the
speech as a delayed event, so it comes after the reading of the title bar.
A call to CancelSpeech furthermore aborts the reading of the title bar.