Implement multiple alarms in accessible WinBoard veersion
Blind players need to be warned at certain times before their
clock runs out. A new persistent string option -alarmTimes can
specify a comma-separated list of times relative to flagging
(in seconds) at which an alarm has to sound. This alarm uses
the sound selected for the 'ICS Alarm' event. When this is set
to 'no sound', the clock times will be spoken.
The alarm sounds on the first clock to reach the specified time.