Implement Duck Chess
An SVG image for a Duck is added. Communication with engine uses the
'Alien Edition' protocol, where a non-final leg is printed by the
engine on a separate line, suffixed by a comma. Peculiarity is that the
destination of the FIDE piece has to be written in place of the old
Duck location, and that the promotion suffix has to be on the final leg.
SAN writes the Duck destination behind the move, separated from it by a
comma. In this case the promotion suffix is on the first leg.